Login section - finished

+++ src/main/java/com/pm/onlinetest/repository/QuestionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.pm.onlinetest.domain.Question;
 
 @Repository
-public interface QuestionRepository extends CrudRepository<Question, Long>{
+public interface QuestionRepository extends CrudRepository<Question, Integer>{
 	public Question findById(Integer id );
 	//public void save(Question q);
 	//public void addQuestion(Question q);
